+++ src/operations/InlayHints.cpp
@@ -108,7 +108,8 @@ struct InlayHintVisitor : public Luau::AstVisitor
                     return true;
 
                 // Skip first item if it is self
-                if (ftv->hasSelf)
+                // TODO: hasSelf is not always specified, so we manually check for the "self" name (https://github.com/Roblox/luau/issues/551)
+                if (ftv->hasSelf || func->args.data[0]->name == "self")
                     it++;
 
                 for (auto param : func->args)
@@ -169,7 +170,9 @@ struct InlayHintVisitor : public Luau::AstVisitor
             for (auto param : call->args)
             {
                 // Skip first item if it is self
-                if (idx == 0 && ftv->hasSelf && call->self)
+                // TODO: hasSelf is not always specified, so we manually check for the "self" name (https://github.com/Roblox/luau/issues/551)
+                if (idx == 0 && (ftv->hasSelf || (namesIt != ftv->argNames.end() && namesIt->has_value() && namesIt->value().name == "self")) &&
+                    call->self)
                     namesIt++;
 
                 if (namesIt == ftv->argNames.end())